New approach to generate multi number image
Each epoch is responsible for training the model in a particular number. If we have the selected numbers as 0 and 5, at the first epoch, the model will only train on the images from the dataset that contains the number 0. The next epoch will handle the images that contain the number 5 and so on.